# frontend/docs/swagger/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ApiResponseMyProfileResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MyProfileResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponsePublicProfileResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PublicProfileResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "BoundaryCoordinateRequest": {
        "properties": {
          "latitude": {
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0,
            "type": "number"
          },
          "longitude": {
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0,
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "type": "object"
      },
...
      "CropProfileResponse": {
        "properties": {
          "cropId": {
            "format": "uuid",
            "type": "string"
          },
          "cropName": {
            "type": "string"
          }
        },
        "required": [
          "cropId",
          "cropName"
        ],
        "type": "object"
      },
...
      "DataSourceRequest": {
        "properties": {
          "address": {
            "type": "string"
          },
          "coordinate": {
            "type": "string"
          },
          "landCharacteristic": {
            "type": "string"
          },
          "parcel": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "FarmRequest": {
        "properties": {
          "areaIsManualEntry": {
            "type": "boolean"
          },
          "areaSqm": {
            "minimum": 0.0,
            "type": "number"
          },
          "boundaryCoordinates": {
            "items": {
              "$ref": "#/components/schemas/BoundaryCoordinateRequest"
            },
            "type": "array"
          },
          "cropIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "dataSource": {
            "$ref": "#/components/schemas/DataSourceRequest"
          },
          "farmId": {
            "format": "uuid",
            "type": "string"
          },
          "jibunAddress": {
            "type": "string"
          },
          "landCategory": {
            "type": "string"
          },
          "latitude": {
            "format": "double",
            "type": "number"
          },
          "longitude": {
            "format": "double",
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "pnu": {
            "type": "string"
          },
          "roadAddress": {
            "type": "string"
          }
        },
        "required": [
          "areaIsManualEntry",
          "boundaryCoordinates",
          "cropIds",
          "dataSource",
          "latitude",
          "longitude",
          "name",
          "roadAddress"
        ],
        "type": "object"
      },
...
      "MyFarmResponse": {
        "properties": {
          "displayRegion": {
            "type": "string"
          },
          "farmId": {
            "format": "uuid",
            "type": "string"
          },
          "jibunAddress": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "roadAddress": {
            "type": "string"
          }
        },
        "required": [
          "farmId",
          "name",
          "roadAddress"
        ],
        "type": "object"
      },
      "MyProfileResponse": {
        "properties": {
          "birthDate": {
            "format": "date",
            "type": "string"
          },
          "crops": {
            "items": {
              "$ref": "#/components/schemas/CropProfileResponse"
            },
            "type": "array"
          },
          "email": {
            "type": "string"
          },
          "experienceLevel": {
            "format": "int32",
            "type": "integer"
          },
          "farms": {
            "items": {
              "$ref": "#/components/schemas/MyFarmResponse"
            },
            "type": "array"
          },
          "managementType": {
            "type": "string"
          },
          "memberId": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profileImageUrl": {
            "type": "string"
          }
        },
        "required": [
          "crops",
          "farms",
          "memberId"
        ],
        "type": "object"
      },
...
      "PublicFarmResponse": {
        "properties": {
          "displayRegion": {
            "type": "string"
          },
          "farmId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "farmId"
        ],
        "type": "object"
      },
      "PublicProfileResponse": {
        "properties": {
          "crops": {
            "items": {
              "$ref": "#/components/schemas/CropProfileResponse"
            },
            "type": "array"
          },
          "experienceLevel": {
            "format": "int32",
            "type": "integer"
          },
          "farms": {
            "items": {
              "$ref": "#/components/schemas/PublicFarmResponse"
            },
            "type": "array"
          },
          "managementType": {
            "type": "string"
          },
          "memberId": {
            "format": "uuid",
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "profileImageUrl": {
            "type": "string"
          }
        },
        "required": [
          "crops",
          "farms",
          "memberId"
        ],
        "type": "object"
      },
...
      "UpdateMyProfileRequest": {
        "properties": {
          "birthDate": {
            "format": "date",
            "type": "string"
          },
          "experienceLevel": {
            "format": "int32",
            "maximum": 100,
            "minimum": 0,
            "type": "integer"
          },
          "farms": {
            "items": {
              "$ref": "#/components/schemas/FarmRequest"
            },
            "type": "array"
          },
          "managementType": {
            "enum": [
              "AGRICULTURAL_INDIVIDUAL",
              "AGRICULTURAL_CORPORATION",
              "NON_REGISTERED_FARMER"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profileMediaId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "birthDate",
          "experienceLevel",
          "farms",
          "managementType",
          "name",
          "nickname",
          "phone"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "access-token": {
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "ChamChamCham Backend API Documentation",
    "title": "ChamChamCham API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/members/me/profile": {
      "put": {
        "operationId": "updateMyProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMyProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMyProfileResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "member-controller"
        ]
      }
    },
    "/api/v1/members/{memberId}/profile": {
      "get": {
        "operationId": "getPublicProfile",
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePublicProfileResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "member-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "access-token": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ]
}